refactor: split logic into separate file
